+++ src/screens/resume-process/achievement/Achievement.css
@@ -3,7 +3,9 @@ html {
     font-size: 16px;
 }
 
-*, *:before, *:after {
+*,
+*:before,
+*:after {
     box-sizing: inherit;
 }
 
@@ -12,7 +14,7 @@ body {
     padding: 0;
     margin: 0;
     font-family: 'Inter', sans-serif;
-    background-color: #F8F8F8;
+    background-color: #F8F8F8 !important;
 }
 
 .content {
@@ -105,7 +107,8 @@ body {
         font-size: 1rem;
     }
 
-    .next, .previous {
+    .next,
+    .previous {
         font-size: 1.2rem;
     }
 }
@@ -115,7 +118,8 @@ body {
         font-size: 0.8rem;
     }
 
-    .next, .previous {
+    .next,
+    .previous {
         font-size: 1rem;
     }
 }
@@ -135,7 +139,8 @@ body {
         font-size: 0.6rem;
     }
 
-    .next, .previous {
+    .next,
+    .previous {
         font-size: 0.8rem;
     }
 }
