+++ src/screens/resume-process/skills/Skill.css
@@ -3,7 +3,9 @@ html {
     font-size: 16px;
 }
 
-*, *:before, *:after {
+*,
+*:before,
+*:after {
     box-sizing: inherit;
 }
 
@@ -95,7 +97,8 @@ body {
         font-size: 1rem;
     }
 
-    .next, .previous {
+    .next,
+    .previous {
         font-size: 1.2rem;
     }
 }
@@ -105,7 +108,8 @@ body {
         font-size: 0.8rem;
     }
 
-    .next, .previous {
+    .next,
+    .previous {
         font-size: 1rem;
     }
 }
@@ -125,7 +129,8 @@ body {
         font-size: 0.6rem;
     }
 
-    .next, .previous {
+    .next,
+    .previous {
         font-size: 0.8rem;
     }
 }
